pipeline/pipeline_requirements.txt
pytest
pylint
requests
pandas
python-dotenv